requirements.txt
ConfigArgParse==1.5.3
easydict==1.9
einops==0.6.0
#h5py==3.6.0
imageio==2.9.0
jaxtyping==0.2.14
lpips==0.1.4
matplotlib==3.5.1
numpy
PyYAML==6.0
roma==1.2.6
scikit_image==0.18.3
scipy
scikit-image
timm==0.5.4
torch
torchvision
tqdm==4.62.3
typeguard==2.13.3
typing_extensions==4.5.0
tensorboard
opencv-python
